requirements.txt
# Core LLM & LangChain
langchain==0.2.16
langchain-community==0.2.16
langchain-groq==0.1.9
langserve[all]==0.2.3
fastapi==0.111.1
uvicorn[standard]==0.30.1
sse-starlette==2.1.3
httpx==0.27.0

# Embeddings & Vector Store
sentence-transformers==3.0.1
faiss-cpu==1.8.0
langchain-huggingface==0.0.3

# Document Parsing
pymupdf==1.24.9
pdfplumber==0.11.4
python-docx==1.1.2

# UI
gradio==4.42.0

# Evaluation
ragas==0.1.14
datasets==2.20.0

# Guardrails
nemoguardrails==0.9.1.1

# Utilities
python-dotenv==1.0.1
pydantic==2.8.2
tiktoken==0.7.0
numpy==1.26.4
pandas==2.2.2
tqdm==4.66.5
